更新模板

+++ template.cpp
@@ -25,7 +25,7 @@ using namespace std; // only in algorithm contests
 
 typedef int64_t ll;
 typedef uint64_t ull;
-typedef double f;
+typedef long double f;
 
 const int dx[4] = {-1,0,1,0};
 const int dy[4] = {0,1,0,-1};
